requirements.txt
streamlit
torch
transformers
pillow
PyMuPDF
colpali-engine